Move vehicle class concatenation to Pandas
Since we are manipulating the retrieved data in pandas it makes sense
to retrieve all of it into an array and concatenate vehicle classes
in Pandas rather than doing it as the data is read in.

Apart from anything else, this makes it easier to try manipulating the
data in different ways.