Explicitly scale the x-axis
Previously the shared x-axis was only owrking because we had data
for the same time ranges for all links. But that's not guaranteed,
and recently some links have stooped producing any data resulting
in inconsistent scaling between pages and data on some graphs running
off the right-hand end.

This edit fixes the problem by explicitly scaling every graph to have
an x-axis that runs from the easrlest to the latest reading in the entire
data set.