Add two experimental scripts to merge locations and journey times
These two scripts use different techniques to merge the data from the
locations and live journey endpoints into a single file. Such files
would be suitable for archiving since they are immune to changes in
the location data. The downside is that the merged files are much larger
that needed just to save the live journey endpoint data.

flatten.py creates a new layout containing a single list of all the links
and compound routes annotated with journey data and with references to
sites expanded into fill site data.

flatten2.py just copies the locations data and annotates all the links and
compoundRoutes with journey time data.